+++ src/provesid/__init__.py
@@ -1,5 +1,5 @@
 # chemid __init__.py
-__version__ = "0.6.0"
+__version__ = "0.7.0"
 
 from .cascommonchem import CASCommonChem
 from .chebi import ChEBI, ChEBIError, ChebiSDF, get_chebi_entity, search_chebi